# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "HRM Platform API",
        "description": "API documentation for HRM Platform",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/recruitment/jobs": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Get all job postings",
                "operationId": "c9a661afc0a77ed20dec12b697fc3fbf",
                "responses": {
                    "200": {
                        "description": "Job postings list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Create a job posting",
                "operationId": "062f16fa01c3082e1dee2bfb3793be29",
                "responses": {
                    "200": {
                        "description": "Job posting created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/recruitment/jobs/{id}": {
            "put": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Update a job posting",
                "operationId": "aab737f363fc0aa080253bdf547bb008",
                "responses": {
                    "200": {
                        "description": "Job posting updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/recruitment/candidates": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Get all candidates",
                "operationId": "be6e6f141d71156fbe55a72ae6b7ee7d",
                "responses": {
                    "200": {
                        "description": "Candidates list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Create a candidate",
                "operationId": "85e2ca6e5f6f6c06f734ed9dbd0b65fe",
                "responses": {
                    "200": {
                        "description": "Candidate created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/recruitment/applications": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Get all applications",
                "operationId": "8436a88c2a94996763ba6cff20c13761",
                "responses": {
                    "200": {
                        "description": "Applications list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Create an application",
                "operationId": "57dbf45dd273d6385252538a647117c4",
                "responses": {
                    "200": {
                        "description": "Application created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/recruitment/applications/{id}": {
            "put": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Update application (move stage, rate, etc.)",
                "operationId": "14560c3b1443a1e04a293feb6ed32179",
                "responses": {
                    "200": {
                        "description": "Application updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/recruitment/summary": {
            "get": {
                "tags": [
                    "Recruitment"
                ],
                "summary": "Get recruitment dashboard summary",
                "operationId": "995540cd072c04e2a3bf11dba66e0747",
                "responses": {
                    "200": {
                        "description": "Recruitment summary"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/employees": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get employee demographics and department distribution",
                "operationId": "cdeb74bf2ded4013b050b2a3057baaea",
                "responses": {
                    "200": {
                        "description": "Employee reports data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/attendance": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get attendance trends",
                "operationId": "32ed80edce3b349ded365c8acb8f7991",
                "responses": {
                    "200": {
                        "description": "Attendance reports data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/payroll": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get payroll expense analytics",
                "operationId": "76da8c0fd42f0d95d5470c11ede957e1",
                "responses": {
                    "200": {
                        "description": "Payroll reports data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/reports/recruitment": {
            "get": {
                "tags": [
                    "Reports"
                ],
                "summary": "Get recruitment and ATS metrics",
                "operationId": "326762882bcac3c72ec9433800e2bc36",
                "responses": {
                    "200": {
                        "description": "Recruitment reports data"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}